implemented react router and a basic navbar for #17
