+++ frontend/package.json
@@ -18,6 +18,7 @@
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
     "react-leaflet": "^3.0.2",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "tailwindcss": "^1.4.6"
   },
